Make separate blueprint for events and orders
